test(web): simplify makes list tests
